Fixed NumericTextField validate()
Previously it would not accept a dot in the text. This was
due to the validate() only receiving one character at a time.
Now we give it the entire text and save the previous,
so in case it isn't a numeric, then the text is restored.

We also restore the caret position on error.

Now we also disallow spaces in a NumericTextField.

Also removed replaceSelection(), since it calls replaceText() any way.
So we only need to override a single function.